Only run install.php if LocalSettings.php doesn't exist

+++ wikibase/1.29/entrypoint.sh
@@ -9,17 +9,18 @@ set -eu
 sleep 1
 /wait-for-it.sh $DB_SERVER -t 60
 
-# Do the mediawiki install
-# TODO if LocalSettings already exists DONT run install.php!!
-php /var/www/html/maintenance/install.php --dbuser $DB_USER --dbpass $DB_PASS --dbname $DB_NAME --dbserver $DB_SERVER --lang $MW_SITE_LANG --pass $MW_ADMIN_PASS $MW_SITE_NAME $MW_ADMIN_NAME
+# Do the mediawiki install (only if LocalSettings doesn't already exist)
+if [ ! -e "/var/www/html/LocalSettings.php" ]; then
+    php /var/www/html/maintenance/install.php --dbuser $DB_USER --dbpass $DB_PASS --dbname $DB_NAME --dbserver $DB_SERVER --lang $MW_SITE_LANG --pass $MW_ADMIN_PASS $MW_SITE_NAME $MW_ADMIN_NAME
 
-# Copy our LocalSettings into place after install from the template
-# https://stackoverflow.com/a/24964089/4746236
-export DOLLAR='$'
-envsubst < /LocalSettings.php.template > /var/www/html/LocalSettings.php
+    # Copy our LocalSettings into place after install from the template
+    # https://stackoverflow.com/a/24964089/4746236
+    export DOLLAR='$'
+    envsubst < /LocalSettings.php.template > /var/www/html/LocalSettings.php
 
-# Run update.php to install Wikibase
-php /var/www/html/maintenance/update.php --quick
+    # Run update.php to install Wikibase
+    php /var/www/html/maintenance/update.php --quick
+fi
 
 # Run the actual entry point
 docker-php-entrypoint apache2-foreground
